address DeepSource and rename embedded_manifold.py to base.py
