+++ geomstats/geometry/matrices.py
@@ -5,7 +5,7 @@ from functools import reduce
 import geomstats.backend as gs
 import geomstats.errors
 from geomstats.algebra_utils import from_vector_to_diagonal_matrix
-from geomstats.geometry.embedded_manifold import VectorSpace
+from geomstats.geometry.base import VectorSpace
 from geomstats.geometry.euclidean import EuclideanMetric
