+++ geomstats/geometry/spd_matrices.py
@@ -4,7 +4,7 @@ import math
 
 import geomstats.backend as gs
 import geomstats.vectorization
-from geomstats.geometry.embedded_manifold import OpenSet
+from geomstats.geometry.base import OpenSet
 from geomstats.geometry.general_linear import GeneralLinear
 from geomstats.geometry.matrices import Matrices
 from geomstats.geometry.riemannian_metric import RiemannianMetric
@@ -353,51 +353,51 @@ class SPDMatrices(SymmetricMatrices, OpenSet):
         logm = gs.reshape(logm, mat.shape)
         return logm
 
-    def is_tangent(self, vector, base_point=None, atol=gs.atol):
-        """Check whether the vector is tangent at base_point.
-
-        A "vector" is tangent to the manifold of SPD matrices if it is a
-        symmetric matrix.
-
-        Parameters
-        ----------
-        vector : array-like, shape=[..., n, n]
-            Matrix.
-        base_point : array-like, shape=[..., n, n]
-            Point on the manifold.
-            Optional, default: None.
-        atol : float
-            Absolute tolerance.
-            Optional, default: backend atol.
-
-        Returns
-        -------
-        is_tangent : bool
-            Boolean denoting if vector is a tangent vector at the base point.
-        """
-        return super(SPDMatrices, self).belongs(vector, atol)
-
-    def to_tangent(self, vector, base_point=None):
-        """Project a vector to a tangent space of the manifold.
-
-        A "vector" is tangent to the manifold of SPD matrices if it is a
-        symmetric matrix, so the symmetric component of the input matrix is
-        returned.
-
-        Parameters
-        ----------
-        vector : array-like, shape=[..., n, n]
-            Matrix.
-        base_point : array-like, shape=[..., n, n]
-            Point on the manifold.
-            Optional, default: None.
-
-        Returns
-        -------
-        tangent_vec : array-like, shape=[..., n, n]
-            Tangent vector at base point.
-        """
-        return super(SPDMatrices, self).projection(vector)
+    # def is_tangent(self, vector, base_point=None, atol=gs.atol):
+    #     """Check whether the vector is tangent at base_point.
+    #
+    #     A "vector" is tangent to the manifold of SPD matrices if it is a
+    #     symmetric matrix.
+    #
+    #     Parameters
+    #     ----------
+    #     vector : array-like, shape=[..., n, n]
+    #         Matrix.
+    #     base_point : array-like, shape=[..., n, n]
+    #         Point on the manifold.
+    #         Optional, default: None.
+    #     atol : float
+    #         Absolute tolerance.
+    #         Optional, default: backend atol.
+    #
+    #     Returns
+    #     -------
+    #     is_tangent : bool
+    #         Boolean denoting if vector is a tangent vector at the base point.
+    #     """
+    #     return super(SPDMatrices, self).belongs(vector, atol)
+    #
+    # def to_tangent(self, vector, base_point=None):
+    #     """Project a vector to a tangent space of the manifold.
+    #
+    #     A "vector" is tangent to the manifold of SPD matrices if it is a
+    #     symmetric matrix, so the symmetric component of the input matrix is
+    #     returned.
+    #
+    #     Parameters
+    #     ----------
+    #     vector : array-like, shape=[..., n, n]
+    #         Matrix.
+    #     base_point : array-like, shape=[..., n, n]
+    #         Point on the manifold.
+    #         Optional, default: None.
+    #
+    #     Returns
+    #     -------
+    #     tangent_vec : array-like, shape=[..., n, n]
+    #         Tangent vector at base point.
+    #     """
+    #     return super(SPDMatrices, self).projection(vector)
 
 
 class SPDMetricAffine(RiemannianMetric):
