--- tests/tests_geomstats/test_lie_group.py
@@ -1,10 +1,0 @@
-"""Unit tests for Lie groups."""
-
-import geomstats.tests
-from geomstats.geometry.lie_group import LieGroup
-
-
-class TestLieGroup(geomstats.tests.TestCase):
-
-    def test_dimension(self):
-        self.assertRaises(TypeError, lambda: LieGroup(4))
